requirements.txt
Django==1.9
djangorestframework==3.4.7
# django-rest-swagger
django-oauth-toolkit